Allow space for a bootloader installed with dd
Implement bootoffset to explcitly allow space for
a bootloader to be installed in front of the
first partition. Add support for cubietruck in
the example to dd the Debian uboot image by adding
a parameter to the customise script call.
Increase size of boot partition to make upgrades easier.
Allow the partition type to be specified.
Use mib units as these are converted to 1024 instead of 1000.